👌 Clean up structs from patch functionality
